+++ src/main/java/lapr/project/model/Client.java
@@ -3,6 +3,7 @@ package lapr.project.model;
 
 import lapr.project.data.ClientDataHandler;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 public class Client extends User{
@@ -13,9 +14,9 @@ public class Client extends User{
     private int numCredits;
     private double latitude;
     private double longitude;
-    private long creditCardNumber;
+    private BigDecimal creditCardNumber;
 
-    public Client(int idClient, String name, String email, String pwd, double nif, double latitude, double longitude,  long creditCardNumber) {
+    public Client(int idClient, String name, String email, String pwd, double nif, double latitude, double longitude,  BigDecimal creditCardNumber) {
         super(email,pwd, "CLIENT");
         this.idClient = idClient;
         this.name = name;
@@ -28,7 +29,7 @@ public class Client extends User{
     }
 
 
-    public Client(String name, String email, String pwd  ,double nif, double latitude, double longitude, long creditCardNumber) {
+    public Client(String name, String email, String pwd  ,double nif, double latitude, double longitude, BigDecimal creditCardNumber) {
         super(email, pwd, "CLIENT");
         this.name = name;
         this.email = email;
@@ -47,7 +48,7 @@ public class Client extends User{
         this.longitude = longitude;
     }
 
-    public Client(String email, String role, int idClient, String name, double nif, double latitude, double longitude, long creditCardNumber, int numCredits) {
+    public Client(String email, String role, int idClient, String name, double nif, double latitude, double longitude, BigDecimal creditCardNumber, int numCredits) {
         super(email, role);
         this.idClient = idClient;
         this.name = name;
@@ -118,11 +119,11 @@ public class Client extends User{
         this.longitude = longitude;
     }
 
-    public long getCreditCardNumber() {
+    public BigDecimal getCreditCardNumber() {
         return creditCardNumber;
     }
 
-    public void setCreditCardNumber(long creditCardNumber) {
+    public void setCreditCardNumber(BigDecimal creditCardNumber) {
         this.creditCardNumber = creditCardNumber;
     }
